+++ sandbox/modules/identity/src/main/java/org/opensearch/identity/authz/Permission.java
@@ -1,9 +1,12 @@
 /*
- * Copyright OpenSearch Contributors
  * SPDX-License-Identifier: Apache-2.0
+ *
+ * The OpenSearch Contributors require contributions made to
+ * this file be licensed under the Apache-2.0 license or a
+ * compatible open source license.
  */
 
-package org.opensearch.authn;
+package org.opensearch.identity.authz;
 
 /**
  * A class that defines a Permission.
